Make the option parsing able to change any config
As long as the config is present, the option parsing will accept an
optional parameter to change it. There is no longer any hard coded
list.

Config, Help and Version still have special handling.

This should make it possible to have one set of parameters for IRC and
another for Discord, while keeping the same option parsing.